feat(adv-prog): Initialise bank account
The bank account class is initialised according to the instructions in
the brief, prompting the user for inputs rather than taking them as
arguments. Ideally, the account number would be randomly generated but
the brief doesn't require that.

Outputs for financial values are formatted such that they look like
normal monetary values e.g. $140.20.